[Feat] 37 SettingWrittenProtocol 생성
